+++ packages/error/src/index.ts
@@ -53,7 +53,7 @@ export function newMessage(err: any, newMsg: string): ErrorWithData {
   const newErr = (isError(err) ? err : new Error(String(err))) as ErrorWithData;
   newErr.data = newErr.data || {};
   newErr.data._previousMessages = newErr.data._previousMessages || [];
-  newErr.data._previousMessages.splice(0, 0, newErr.message);
+  newErr.data._previousMessages.unshift(newErr.message);
   newErr.message = `${newMsg}: ${newErr.message}`;
   return newErr;
 }